Add new dotfiles; update 'install'
.bash_profile and .profile have been added.
Considering .bashrc is meant for more of an interactive
non-login bash session, I chose to follow this convention
and keep it so.

'install' has been edited to work with '.profile'.
This is mainly to follow suit from above, but also
that interactive login sessions should be able to also
have access to the same env vars currently.